\name{ambur-package}
\alias{ambur-package}
\alias{ambur}
\docType{package}
\title{
Analyzing Moving Boundaries Using R - AMBUR
}
\description{
Performs an analysis on position changes of a boundary (shoreline) along transects. Mainly used for shoreline change analysis.

}
\details{
\tabular{ll}{
Package: \tab ambur\cr
Type: \tab Package\cr
Version: \tab 1.03-12\cr
Date: \tab 2011-08-17\cr
License: \tab GNU General Public License (please site references below)\cr
LazyLoad: \tab yes\cr
}
This package works with data stored in *DBF, *CSV, and ESRI Shapefile format depending on the function.
}
\author{
Chester W. Jackson Jr.\cr

Maintainer:\cr 
Chester W. Jackson Jr.\cr 
\email{jackson.cwjr@gmail.com}\cr
Department of Geology and Georgraphy\cr
Georgia Southern University\cr
}
\references{
JACKSON, C.W., 2009. Assessing Historical Shoreline Changes And Erosion Hazards Using Tools Developed For ArcGIS and R , NOAA Coastal GeoTools 2009 Conference, March 2-5, 2009, Myrtle Beach, SC.  \cr
\cr
JACKSON, C.W., 2010. SPATIO-TEMPORAL ANALYSIS OF BARRIER ISLAND SHORELINE CHANGE: THE GEORGIA COAST, USA.  Ph.D. Dissertation, University of Georgia.\cr
}

\keyword{ ambur }
\seealso{ http://ambur.r-forge.r-project.org/    }


