\name{ambur.gui}
\alias{ambur.gui}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ ambur's Graphical User Interface }
\description{
Provides access to functions via windows and buttons
}
\usage{
ambur.gui()
}

\details{
type 'ambur.gui()' at the command prompt to start the GUI
}
\value{
n/a
}
\references{ }
\note{ }
\seealso{  }
\examples{
To start the ambur GUI:

ambur.gui()

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ gui }
\keyword{ interface }% __ONLY ONE__ keyword per line
