\name{ambur.statshape}
\alias{ambur.statshape}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Construct shapefiles from AMBUR analysis stats file }
\description{
Construct shapefiles to help visualize shoreline change statistics in GIS.
}
\usage{
ambur.statshape( )
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{ }{  }
}
\details{
Builds shapefiles using the 'result_stats.csv' file produced from the ambur.analysis function.
}
\value{

Shapefiles with attribute tables containing statistics from the 'result_stats.csv' file.
}
\references{ }
\note{}
\seealso{  }
\examples{
-to build the shapefiles, type:
ambur.statshape()


}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ stat }
\keyword{ shapefile }% __ONLY ONE__ keyword per line